+++ src/editor/i18n.ts
@@ -1,10 +1,12 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2020 The Pybricks Authors
+// Copyright (c) 2020-2021 The Pybricks Authors
 //
 // Editor translation keys.
 
 export enum EditorStringId {
     Placeholder = 'editor.placeholder',
+    Check = 'editor.check',
+    ToggleDocs = 'editor.toggleDocs',
     Copy = 'editor.copy',
     Paste = 'editor.paste',
     SelectAll = 'editor.selectAll',