+++ src/react-app-env.d.ts
@@ -14,3 +14,8 @@ declare module '*.zip' {
     const src: string;
     export default src;
 }
+
+// https://webpack.js.org/api/hot-module-replacement/
+interface NodeModule {
+    hot?: { dispose: (callback: (data) => void) => void };
+}